% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate confidence intervals}
\usage{
calc_ci(point, se, alpha = 0.05)
}
\arguments{
\item{point}{point estimate}

\item{se}{standard error of point estimate}

\item{alpha}{alpha value for confidence interval (defaul to .05)}
}
\value{
list with two elements, $lb and $ub, which are the lower and 
upper bounds of the confidence interval respectively
}
\description{
calc_ci() is a function that calculates a confidence interval using the 
normal distribution given a point estimate, SE, and alpha value
}
\examples{
calc_ci

}
