% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_stats.R
\name{calc_sum_stats}
\alias{calc_sum_stats}
\title{Calculate sample summary statistics}
\usage{
calc_sum_stats(DT, sum_vars, by_var = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{sum_vars}{A vector of string column names in DT to calculate summary
statistics of.}

\item{by_var}{A string of the column name in DT to group summary statistics
by}
}
\value{
data.table of summary statistics
}
\description{
calc_sum_stats() is a funtion that calculates summary statistics (mean,
standard deviation, minimum, and maximum) for the either the entire sample,
by group, or both
}
\examples{
# mtcars data set
DT <- as.data.table(mtcars)
# entire sample
calc_sum_stats(DT, sum_vars = c("mpg", "hp"))

}
