% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_log_file.R
\name{end_log_file}
\alias{end_log_file}
\title{End a log file}
\usage{
end_log_file(log_file)
}
\arguments{
\item{log_file}{if TRUE output/messages are being diverted to a .log
file and end_log_file() will close the connection}
}
\description{
end_log_file() is a function that closes the connection opened by
start_log_file()
}
