% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Removes observations with NAs}
\usage{
remove_na(DT, cols, track = F)
}
\arguments{
\item{DT}{a data.table}

\item{cols}{columns to look for NA observations}
}
\value{
The data.table without NAs in any of the specified columns
}
\description{
remove_na is a function that removes observations if any specified column is 
NA
}
\examples{
# 2013 nyc flights data
DT <- as.data.table(nycflights13::flights)

DT_clean <- remove_na(DT = DT, cols = c("arr_delay", "carrier"))
}
