% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_log_file.R
\name{start_log_file}
\alias{start_log_file}
\title{Start a log file}
\usage{
start_log_file(file_name)
}
\arguments{
\item{file_name}{path to .log file where diverted output and
messages will be written}
}
\description{
start_log_file() is a function that opens a connection and diverts output/
messages to a .log file
}
