% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cmean.R
\name{calc_cmean}
\alias{calc_cmean}
\title{Calculate conditional sample mean/SEs}
\usage{
calc_cmean(DT, y, x, se = F, alpha = 0.05)
}
\arguments{
\item{DT}{a data.table}

\item{y}{a character vector of column names of DT to calculate the group mean
of}

\item{x}{a character vector of columns in DT to group by}

\item{se}{if TRUE output will include standard error of conditional sample
mean}

\item{alpha}{a real number between 0 and 1, to use as the alpha level in
confidence interval construction (default is .05)}
}
\value{
The data.table of conditional sample means
}
\description{
calc_cmean() is a function that calculates sample means by group, their
standard errors, and confidence intervals
}
\examples{
DT <- data.table(x1 = sample(1:5, 1000, replace = T), 
                 x2 = sample(1:2, 1000, replace = T)) \%>\% 
      .[, y := 2*x1 + 4*x2 + rnorm(mean = 10, sd = 4, n = 1000)]

calc_cmean(DT, y = c("y1"), x = c("x1", "x2"))

}
